Basic classes added

+++ forger-workflow/src/main/kotlin/com/github/wenslo/forger/workflow/domain/ExecuteShip.kt
@@ -5,7 +5,7 @@ package com.github.wenslo.forger.workflow.domain
  * @date 2022/11/1 21:19
  */
 data class ExecuteShip(
-    var playScriptId: Int = 0,
+    var playScriptId: Long = 0,
     var playScriptUniqueId: String = "",
     var playScriptRecordId: Long = 0L,
     var current: String = "",
