+++ forger-workflow/src/main/kotlin/com/github/wenslo/forger/workflow/entity/PlayScriptExecuteRecord.kt
@@ -8,7 +8,7 @@ import com.github.wenslo.forger.workflow.enums.PlayScriptProcessStatus
  * @date 2022/10/27 10:44
  */
 data class PlayScriptExecuteRecord(
-    var playScriptId: Int = 0,
+    var playScriptId: Long = 0,
     var playScriptUniqueId: String = "",
     var processStatus: PlayScriptProcessStatus = PlayScriptProcessStatus.NONE,
     var current: List<String> = emptyList(),
